requirements.txt
# API для криптобирж
ccxt>=2.0.0

# Анализ данных
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Технический анализ
ta>=0.9.0
#ta-lib>=0.4.0  # Опционально, требует отдельной установки

# Машинное обучение
scikit-learn>=1.0.0
tensorflow>=2.8.0  # или PyTorch

# Утилиты
pyyaml>=6.0
python-dotenv>=0.19.0